feat: rewrite logic
