+++ packages/webapp/pages/plus/index.tsx
@@ -5,10 +5,8 @@ import dynamic from 'next/dynamic';
 import { useViewSize, ViewSize } from '@dailydotdev/shared/src/hooks';
 import { useRouter } from 'next/router';
 import type { NextSeoProps } from 'next-seo/lib/types';
-import { getUserShortInfo } from '@dailydotdev/shared/src/graphql/users';
-import type { GetStaticPropsContext, GetStaticPropsResult } from 'next';
-import type { ParsedUrlQuery } from 'querystring';
-import type { EmptyObjectLiteral } from '@dailydotdev/shared/src/lib/kratos';
+import type { GiftUserContextData } from '@dailydotdev/shared/src/components/plus/GiftUserContext';
+import { GiftUserContext } from '@dailydotdev/shared/src/components/plus/GiftUserContext';
 import { getPlusLayout } from '../../components/layouts/PlusLayout/PlusLayout';
 import { getTemplatedTitle } from '../../components/layouts/utils';
 import { defaultOpenGraph } from '../../next-seo';
@@ -31,7 +29,9 @@ const seo: NextSeoProps = {
     'Upgrade to daily.dev Plus for an ad-free experience, custom feeds, bookmark folders, clickbait shield, and more.',
 };
 
-const PlusPage = (): ReactElement => {
+export type PlusPageProps = Pick<GiftUserContextData, 'giftToUser'>;
+
+const PlusPage = ({ giftToUser }: PlusPageProps): ReactElement => {
   const { isReady } = useRouter();
   const isLaptop = useViewSize(ViewSize.Laptop);
 
@@ -39,45 +39,14 @@ const PlusPage = (): ReactElement => {
     return null;
   }
 
-  if (isLaptop) {
-    return <PlusDesktop />;
-  }
-
-  return <PlusMobile />;
+  return (
+    <GiftUserContext.Provider value={{ giftToUser }}>
+      {isLaptop ? <PlusDesktop /> : <PlusMobile />}
+    </GiftUserContext.Provider>
+  );
 };
 
 PlusPage.getLayout = getPlusLayout;
 PlusPage.layoutProps = { seo };
-
-interface PageProps extends ParsedUrlQuery {
-  giftToUserId: string;
-}
-
-export const getStaticProps = async ({
-  params,
-}: GetStaticPropsContext<PageProps>): Promise<
-  GetStaticPropsResult<EmptyObjectLiteral>
-> => {
-  const validateUserId = (value: string) => !!value && value !== '404';
-  const giftToUserId = params?.giftToUserId;
-
-  if (!validateUserId(giftToUserId)) {
-    return { props: {} };
-  }
-
-  try {
-    const giftToUser = await getUserShortInfo(giftToUserId);
-
-    if (giftToUser.isPlus) {
-      return { props: {} };
-    }
-
-    return {
-      redirect: { destination: `/plus/gift/${giftToUserId}`, permanent: false },
-    };
-  } catch (err) {
-    return { props: {} };
-  }
-};
 
 export default PlusPage;
